+++ pkg/timeDate/R/holiday-Listing.R
@@ -78,8 +78,8 @@ listHolidays <-
         "CHAscension",
         "CHConfederationDay",
         "CHKnabenschiessen",
-        "GBMayDay",
-        "GBBankHoliday",
+        "GBEarlyMayBankHoliday", # was: "GBMayDay",
+        "GBSpringBankHoliday",   # was: "GBBankHoliday"
         "GBSummerBankHoliday",
         "GBMilleniumDay",
         "DEAscension",