+++ pkg/timeDate/inst/unitTests/runit.HolidayCalendars.R
@@ -239,6 +239,69 @@ test.holidayNERC <- function() {
     test()
 }
 
+## GNB: test fixes on 2022-10-01
+test.holidayLONDON <- function() {
+
+    test <- function() {
+        # Holiday LONDON -
+        HD <- holidayLONDON(getRmetricsOptions("currentYear"))
+        print(HD)
+        checkTrue(inherits(HD, "timeDate"))
+
+        ## https://www.gov.uk/bank-holidays#england-and-wales
+        ## (downloaded on 2022-10-01, this is updated regularly
+        
+        ph2017 <- c("2017-01-02", "2017-04-14", "2017-04-17", "2017-05-01",
+                    "2017-05-29", "2017-08-28", "2017-12-25", "2017-12-26" )
+        
+        ph2018 <- c("2018-01-01", "2018-03-30", "2018-04-02", "2018-05-07",
+                    "2018-05-28", "2018-08-27", "2018-12-25", "2018-12-26" )
+
+        ph2019 <- c("2019-01-01", "2019-04-19", "2019-04-22", "2019-05-06",
+                    "2019-05-27", "2019-08-26", "2019-12-25", "2019-12-26" )
+
+        ph2020 <- c("2020-01-01", "2020-04-10", "2020-04-13", "2020-05-08",
+                    "2020-05-25", "2020-08-31", "2020-12-25", "2020-12-28" )
+
+        ph2021 <- c("2021-01-01", "2021-04-02", "2021-04-05", "2021-05-03",
+                    "2021-05-31", "2021-08-30", "2021-12-27", "2021-12-28" )
+
+        ## Queen's Platinum Jubilee; Quinn's death
+        ph2022 <- c("2022-01-03", "2022-04-15", "2022-04-18", "2022-05-02",
+                    "2022-06-02", "2022-06-03", "2022-08-29", "2022-09-19",
+                    "2022-12-26", "2022-12-27" ) 
+
+        ## Queen's Jubilee
+        ph2002 <- c("2002-01-01", "2002-03-29", "2002-04-01", "2002-05-06",
+                    "2002-06-03", "2002-06-04", "2002-08-26", "2002-12-25", "2002-12-26")
+
+        ## Queen's Diamond Jubilee
+        ph2012 <- c("2012-01-02", "2012-04-06", "2012-04-09", "2012-05-07",
+                    "2012-06-04", "2012-06-05", "2012-08-27", "2012-12-25", "2012-12-26")
+
+        checkTrue(all.equal(ph2017, format(holidayLONDON(2017))))
+        checkTrue(all.equal(ph2018, format(holidayLONDON(2018))))
+        checkTrue(all.equal(ph2019, format(holidayLONDON(2019))))
+        checkTrue(all.equal(ph2020, format(holidayLONDON(2020))))
+        checkTrue(all.equal(ph2021, format(holidayLONDON(2021))))
+        checkTrue(all.equal(ph2022, format(holidayLONDON(2022))))
+
+        checkTrue(all.equal(ph2002, format(holidayLONDON(2002))))
+        checkTrue(all.equal(ph2012, format(holidayLONDON(2012))))
+
+        
+        checkTrue(format(GBMayDay(1995)) == "1995-05-08") # VE day that year
+        checkTrue(format(GBMayDay(2020)) == "2020-05-08") # VE day that year
+
+        checkTrue(format(GBBankHoliday(2002)) == "2002-06-03")
+        checkTrue(format(GBBankHoliday(2012)) == "2012-06-04")
+        checkTrue(format(GBBankHoliday(2022)) == "2022-06-02")
+
+    }
+
+    test()
+}
+